+++ echopype/tests/calibrate/test_calibrate.py
@@ -5,7 +5,7 @@ from scipy.io import loadmat
 import echopype as ep
 from echopype.calibrate.calibrate_ek import CalibrateEK80
 from echopype.calibrate.env_params import EnvParams
-from echopype.calibrate.ek80_utils import get_transmit_signal, compress_pulse
+from echopype.calibrate.ek80_complex import get_transmit_signal, compress_pulse
 import xarray as xr
